refactor: studentInfoStore, schoolInfoStore 파일 분리
